+++ src/onegov.gazette/onegov/gazette/layout.py
@@ -177,12 +177,6 @@ class Layout(ChameleonLayout):
 
         return result
 
-    def format_organization(self, organization):
-        return self.principal.organizations.get(organization, '')
-
-    def format_category(self, category):
-        return self.principal.categories.get(category, '')
-
     def format_issue(self, issue, date_format='date'):
         if not isinstance(issue, Issue):
             issue = Issue.from_string(str(issue))