+++ src/onegov.gazette/onegov/gazette/models/notice.py
@@ -1,9 +1,12 @@
 from onegov.chat import Message
 from onegov.core.orm.mixins import meta_property
 from onegov.gazette import _
+from onegov.gazette.models.principal import Issue
 from onegov.notice import OfficialNotice
 from onegov.user import User
 from onegov.user import UserCollection
+from sedate import as_datetime
+from sedate import standardize_date
 from sqlalchemy.orm import backref
 from sqlalchemy.orm import object_session
 from sqlalchemy.orm import relationship
@@ -32,6 +35,14 @@ class GazetteNotice(OfficialNotice):
             self.meta['issues'] = value
         else:
             self.meta['issues'] = {item: None for item in value}
+
+    #: The ID of the organization. We store this in addition to the
+    #: organization name to allow changing organization names.
+    organization_id = meta_property('organization_id')
+
+    #: The ID of the category. We store this in addition to the
+    #: category name to allow changing category names.
+    category_id = meta_property('category_id')
 
     def add_change(self, request, event, text=None):
         """ Adds en entry to the changelog. """
@@ -91,6 +102,25 @@ class GazetteNotice(OfficialNotice):
                 return change.text
 
         return ''
+
+    def apply_meta(self, principal):
+        """ Updates the category, organization and issue date from the meta
+        values.
+
+        """
+        self.organization = principal.organizations.get(self.organization_id)
+        self.category = principal.categories.get(self.category_id)
+
+        issues = [Issue.from_string(issue) for issue in self.issues]
+        issues = [
+            principal.issues.get(issue.year, {}).get(issue.number)
+            for issue in issues
+        ]
+        issues = sorted([issue for issue in issues if issue])
+        if issues:
+            self.issue_date = standardize_date(
+                as_datetime(issues[0]), 'Europe/Zurich'
+            )
 
 
 class GazetteNoticeChange(Message):